회원 관리 기능 일단 마무리
회원 게시글 작성, 삭제 구현

+++ src/main/java/com/sparta/week3_1/service/BoardService.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.util.List;
 
-import static com.sparta.week3_1.ExceptionHandler.ErrorCode.WRONG_PASSWORD;
+import static com.sparta.week3_1.ExceptionHandler.ErrorCode.NO_AUTHORITY;
 
 @Service
 @RequiredArgsConstructor
@@ -29,16 +29,16 @@ public class BoardService {
         );
     }
 
-    public Article writePost(ArticleRequestDto requestDto) {
-        Article article = new Article(requestDto); // 받아온 dto로 엔티티 생성해 db에 저장
+    public Article writePost(ArticleRequestDto requestDto, String author) {
+        Article article = new Article(requestDto, author); // 받아온 dto로 엔티티 생성해 db에 저장
         return boardRepository.save(article);
     }
 
-    public void checkPw(Long id, ArticleRequestDto requestDto) {
-        if (requestDto.getPassword() != getPost(id).getPassword()) {
-            throw new CustomException(WRONG_PASSWORD);
-        }
-    }
+    //public void checkPw(Long id, ArticleRequestDto requestDto) {
+    //    if (requestDto.getPassword() != getPost(id).getPassword()) {
+    //        throw new CustomException(WRONG_PASSWORD);
+    //    }
+    //}
 
     @Transactional // update에는 트랜잭션 필수
     public Article updatePost(Long id, ArticleRequestDto requestDto) {
@@ -49,7 +49,14 @@ public class BoardService {
         return article;
     }
 
-    public void deletePost(Long id) {
-        boardRepository.deleteById(id);
+    public void deletePost(Long id, String author) {
+        Article article = boardRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("해당 아이디가 존재하지 않습니다.")
+        );
+        if (author.equals(article.getAuthor())) {
+            boardRepository.deleteById(id);
+        } else {
+            throw new CustomException(NO_AUTHORITY);
+        }
     }
 }
